# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseDetail": {
        "properties": {
          "assignment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssignmentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "capabilities": {
            "$ref": "#/components/schemas/CaseCapabilities",
            "description": "Computed for the caller."
          },
          "case": {
            "$ref": "#/components/schemas/CaseSummary"
          },
          "channelIdentity": {
            "$ref": "#/components/schemas/ChannelIdentity"
          },
          "closure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseClosure"
              },
              {
                "type": "null"
              }
            ]
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerProfile"
          },
          "routing": {
            "$ref": "#/components/schemas/RoutingSummary",
            "description": "\"Cómo llegó a ti\"."
          }
        },
        "required": [
          "case",
          "customer",
          "channelIdentity",
          "assignment",
          "routing",
          "closure",
          "capabilities"
        ],
        "title": "CaseDetail",
        "type": "object"
      },
...
      "CloseCaseRequest": {
        "additionalProperties": false,
        "properties": {
          "contactReason": {
            "$ref": "#/components/schemas/ContactReason"
          },
          "followUp": {
            "$ref": "#/components/schemas/FollowUp"
          },
          "resolutionCode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          },
          "sendCsatSurvey": {
            "title": "Sendcsatsurvey",
            "type": "boolean"
          }
        },
        "required": [
          "resolved",
          "contactReason",
          "resolutionCode",
          "followUp",
          "sendCsatSurvey"
        ],
        "title": "CloseCaseRequest",
        "type": "object"
      },
...
      "InboxResponse": {
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/InboxCounts",
            "description": "Always the whole inbox (ignores status and q)."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "serverTime": {
            "format": "date-time",
            "title": "Servertime",
            "type": "string"
          }
        },
        "required": [
          "items",
          "counts",
          "serverTime"
        ],
        "title": "InboxResponse",
        "type": "object"
      },
      "InboxStatus": {
        "description": "Canvas bucket of a case in \"Casos\" (derived from ``CaseStatus``, never stored).",
        "enum": [
          "new",
          "to_reply",
          "live",
          "to_call",
          "waiting"
        ],
        "title": "InboxStatus",
        "type": "string"
      },
...
      "MarkReadRequest": {
        "additionalProperties": false,
        "properties": {
          "upToSequence": {
            "minimum": 0.0,
            "title": "Uptosequence",
            "type": "integer"
          }
        },
        "required": [
          "upToSequence"
        ],
        "title": "MarkReadRequest",
        "type": "object"
      },
...
      "PostAnalystTurnRequest": {
        "additionalProperties": false,
        "properties": {
          "clientMessageId": {
            "description": "Client-generated id (UUID v4); also the Idempotency-Key header.",
            "examples": [
              "0b8f6a52-6f0e-4c1e-9d55-2f5a0c7d9e10"
            ],
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Clientmessageid",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "clientMessageId"
        ],
        "title": "PostAnalystTurnRequest",
        "type": "object"
      },
...
      "PostTurnResponse": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseSummary"
          },
          "turn": {
            "$ref": "#/components/schemas/Turn"
          }
        },
        "required": [
          "turn",
          "case"
        ],
        "title": "PostTurnResponse",
        "type": "object"
      },
      "ProblemCode": {
        "enum": [
          "invalid_credentials",
          "mfa_invalid",
          "mfa_challenge_invalid",
          "account_locked",
          "unauthenticated",
          "session_expired",
          "forbidden",
          "not_found",
          "method_not_allowed",
          "conflict",
          "concurrent_update",
          "invalid_transition",
          "case_not_assigned",
          "case_closed",
          "channel_not_supported",
          "idempotency_conflict",
          "invalid_value",
          "policy_violation",
          "validation_error",
          "domain_error",
          "application_error",
          "invalid_topic",
          "invalid_message",
          "http_error",
          "internal_error"
        ],
        "title": "ProblemCode",
        "type": "string"
      },
      "ProblemDetails": {
        "additionalProperties": true,
        "description": "RFC 7807 problem. ``code`` is the stable machine identifier clients branch on.\n\nThe optional members below are the documented extensions; a domain error may add other\nstructured details (e.g. ``channel`` on ``channel_not_supported``), hence\n``additionalProperties``.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "currentStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseStatus"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "invalid_transition, case_closed: the case status now."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "validation_error: one entry per invalid field.",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "remainingAttempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "invalid_credentials, mfa_invalid: failed attempts left before the lock.",
            "title": "Remainingattempts"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestid"
          },
          "requiredRoles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StaffRole"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "forbidden: roles that may perform the action.",
            "title": "Requiredroles"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "examples": [
              "urn:cc-platform:problem:account_locked"
            ],
            "title": "Type",
            "type": "string"
          },
          "unlockAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "account_locked: when the lock ends.",
            "title": "Unlockat"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "TurnPage": {
        "properties": {
          "items": {
            "description": "Ascending sequence.",
            "items": {
              "$ref": "#/components/schemas/Turn"
            },
            "title": "Items",
            "type": "array"
          },
          "lastSequence": {
            "title": "Lastsequence",
            "type": "integer"
          },
          "olderCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldercursor"
          }
        },
        "required": [
          "items",
          "olderCursor",
          "lastSequence"
        ],
        "title": "TurnPage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CustomerToken": {
        "description": "Customer token returned by POST /api/v1/customer/sessions (simulator).",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionToken": {
        "description": "Session token returned by POST /api/v1/auth/mfa.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API of the LATAM Bank contact-center platform (transaction-dispute intake). Errors are RFC 7807 `application/problem+json` with a stable `code`. Realtime: WebSocket `/api/v1/ws?token=…` with `{type,id,occurredAt,data}` envelopes.",
    "title": "CC Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases/inbox": {
      "get": {
        "description": "`counts` always cover the whole inbox (the counters are the filters), whatever `status` and `q` select. Sorted: live calls first, then the closest SLA, then the oldest case. At most 200 items.",
        "operationId": "cases_get_inbox",
        "parameters": [
          {
            "description": "Omit for Todos",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InboxStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Omit for Todos",
              "title": "Status"
            }
          },
          {
            "description": "Customer name or case id (contains)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Customer name or case id (contains)",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "The caller's open cases (\"Casos\") and the status counters",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{caseId}": {
      "get": {
        "operationId": "cases_get_case",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "examples": [
                "CASE-01J…"
              ],
              "maxLength": 64,
              "title": "Caseid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807): validation_error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Case detail: customer, channel identity, assignment, \"Cómo llegó a ti\", closure",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{caseId}/close": {
      "post": {
        "operationId": "cases_close_case",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "examples": [
                "CASE-01J…"
              ],
              "maxLength": 64,
              "title": "Caseid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Close the case (contract case_close); the customer is told the chat ended",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{caseId}/read": {
      "post": {
        "operationId": "cases_mark_read",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "examples": [
                "CASE-01J…"
              ],
              "maxLength": 64,
              "title": "Caseid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Move the assignee's read cursor (opening a new case starts it)",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{caseId}/turns": {
      "get": {
        "description": "No params → the latest page. `cursor` (from `olderCursor`) → the page before it. `afterSequence` → turns after it (gap fill after a reconnect). `cursor` and `afterSequence` are mutually exclusive.",
        "operationId": "cases_list_turns",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "examples": [
                "CASE-01J…"
              ],
              "maxLength": 64,
              "title": "Caseid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "afterSequence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aftersequence"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "A page of the transcript (ascending sequence)",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "description": "Idempotent on `clientMessageId` (= `Idempotency-Key`): a retry with the same text answers 200 with `Idempotent-Replayed: true` and the original turn; the same id with another text is `idempotency_conflict`. Moves a `new` case to `in_progress`.",
        "operationId": "cases_post_turn",
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "examples": [
                "CASE-01J…"
              ],
              "maxLength": 64,
              "title": "Caseid",
              "type": "string"
            }
          },
          {
            "description": "Must equal the body `clientMessageId`; a retry with it is a replay.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Must equal the body `clientMessageId`; a retry with it is a replay.",
              "maxLength": 64,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostAnalystTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostTurnResponse"
                }
              }
            },
            "description": "Replay of an already-created turn"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Reply to the customer (assignee only; chat channels only)",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}